Project's estimated & actual duration, Event's & Task's duration now can be measured in different units. Stored in seconds in the db though.

+++ db/migrate/20110311194653_create_projects.rb
@@ -10,8 +10,8 @@ class CreateProjects < ActiveRecord::Migration
       t.date :estimated_end_date
       t.date :started_on
       t.date :ended_on
-      t.integer :estimated_duration
-      t.integer :actual_duration
+      t.float :estimated_duration
+      t.float :actual_duration
 
       t.timestamps
     end
